using a long conversion is not reliable, switched to int
--HG--
extra : convert_revision : svn%3Aa38d40e9-c014-0410-b785-c606c0c8e7de/satchmo/trunk%401293

+++ satchmo/product/comments.py
@@ -85,7 +85,7 @@ def highest_rated(num=None, free=False, site=None):
     
     if pks:
         productdict = Product.objects.in_bulk(pks)
-        products = [productdict[long(pk)] for pk in pks]
+        products = [productdict[int(pk)] for pk in pks]
     else:
         products = []
